refactor: divided litellm chat and embedding
