+++ src/solace_ai_connector/components/general/llm/litellm/litellm_base.py
@@ -14,12 +14,6 @@ litellm_info_base = {
     "description": "Base class for LiteLLM chat models",
     "config_parameters": [
         {
-            "name": "action",
-            "required": True,
-            "description": "The action to perform (e.g., 'inference', 'embedding')",
-            "default": "inference",
-        },
-        {
             "name": "load_balancer",
             "required": False,
             "description": (
@@ -89,60 +83,10 @@ litellm_info_base = {
             "type": "boolean",
         },
     ],
-    "input_schema": {
-        "type": "object",
-        "properties": {
-            "messages": {
-                "type": "array",
-                "items": {
-                    "type": "object",
-                    "properties": {
-                        "role": {
-                            "type": "string",
-                            "enum": ["system", "user", "assistant"],
-                        },
-                        "content": {"type": "string"},
-                    },
-                    "required": ["role", "content"],
-                },
-            },
-        },
-        "required": ["messages"],
-    },
-    "output_schema": {
-        "type": "object",
-        "properties": {
-            "content": {
-                "type": "string",
-                "description": "The generated response from the model",
-            },
-            "chunk": {
-                "type": "string",
-                "description": "The current chunk of the response",
-            },
-            "response_uuid": {
-                "type": "string",
-                "description": "The UUID of the response",
-            },
-            "first_chunk": {
-                "type": "boolean",
-                "description": "Whether this is the first chunk of the response",
-            },
-            "last_chunk": {
-                "type": "boolean",
-                "description": "Whether this is the last chunk of the response",
-            },
-            "streaming": {
-                "type": "boolean",
-                "description": "Whether this is a streaming response",
-            },
-        },
-        "required": ["content"],
-    },
 }
 
 
-class LiteLLMChatModelBase(ComponentBase):
+class LiteLLMBase(ComponentBase):
     def __init__(self, module_info, **kwargs):
         super().__init__(module_info, **kwargs)
         self.init()
@@ -150,7 +94,6 @@ class LiteLLMChatModelBase(ComponentBase):
 
     def init(self):
         litellm.suppress_debug_info = True
-        self.action = self.get_config("action")
         self.load_balancer = self.get_config("load_balancer")
         self.stream_to_flow = self.get_config("stream_to_flow")
         self.stream_to_next_component = self.get_config("stream_to_next_component")
@@ -182,168 +125,4 @@ class LiteLLMChatModelBase(ComponentBase):
 
     def invoke(self, message, data):
         """invoke the model"""
-        messages = data.get("messages", [])
-
-        if self.action == "inference":
-            if self.llm_mode == "stream":
-                return self.invoke_stream(message, messages)
-            else:
-                return self.invoke_non_stream(messages)
-        elif self.action == "embedding":
-            return self.invoke_embedding(messages)
-        else:
-            raise ValueError(f"Unsupported action: {self.action}") 
-
-    def invoke_embedding(self, messages):
-        """invoke the embedding model"""
-        response = self.router.embedding(model=self.load_balancer[0]["model_name"], 
-                input=messages[0]["content"])
-        log.debug("Embedding response: %s", response)
-
-        # Extract the embedding data from the response
-        embedding_data = response['data'][0]['embedding']
-        return {"embedding": embedding_data}
-
-    def invoke_non_stream(self, messages):
-        """invoke the model without streaming"""
-        max_retries = 3
-        while max_retries > 0:
-            try:
-                response = self.load_balance(messages, stream=False)
-                return {"content": response.choices[0].message.content}
-            except Exception as e:
-                    log.error("Error invoking LiteLLM: %s", e)
-                    max_retries -= 1
-                    if max_retries <= 0:
-                        raise e
-                    else:
-                        time.sleep(1)
-
-    def invoke_stream(self, message, messages):
-        """invoke the model with streaming"""
-        response_uuid = str(uuid.uuid4())
-        if self.set_response_uuid_in_user_properties:
-            message.set_data("input.user_properties:response_uuid", response_uuid)
-
-        aggregate_result = ""
-        current_batch = ""
-        first_chunk = True
-
-        max_retries = 3
-        while max_retries > 0:
-            try:
-                response = self.load_balance(messages, stream=True)
-
-                for chunk in response:
-                    # If we get any response, then don't retry
-                    max_retries = 0
-                    if chunk.choices[0].delta.content is not None:
-                        content = chunk.choices[0].delta.content
-                        aggregate_result += content
-                        current_batch += content
-                        if len(current_batch.split()) >= self.stream_batch_size:
-                            if self.stream_to_flow:
-                                self.send_streaming_message(
-                                    message,
-                                    current_batch,
-                                    aggregate_result,
-                                    response_uuid,
-                                    first_chunk,
-                                    False,
-                                )
-                            elif self.stream_to_next_component:
-                                self.send_to_next_component(
-                                    message,
-                                    current_batch,
-                                    aggregate_result,
-                                    response_uuid,
-                                    first_chunk,
-                                    False,
-                                )
-                            current_batch = ""
-                            first_chunk = False
-            except Exception as e:
-                log.error("Error invoking LiteLLM: %s", e)
-                max_retries -= 1
-                if max_retries <= 0:
-                    raise e
-                else:
-                    # Small delay before retrying
-                    time.sleep(1)
-
-        if self.stream_to_next_component:
-            # Just return the last chunk
-            return {
-                "content": aggregate_result,
-                "chunk": current_batch,
-                "response_uuid": response_uuid,
-                "first_chunk": first_chunk,
-                "last_chunk": True,
-                "streaming": True,
-            }
-
-        if self.stream_to_flow:
-            self.send_streaming_message(
-                message,
-                current_batch,
-                aggregate_result,
-                response_uuid,
-                first_chunk,
-                True,
-            )
-
-        return {"content": aggregate_result, "response_uuid": response_uuid}
-
-    def send_streaming_message(
-        self,
-        input_message,
-        chunk,
-        aggregate_result,
-        response_uuid,
-        first_chunk=False,
-        last_chunk=False,
-    ):
-        message = Message(
-            payload={
-                "chunk": chunk,
-                "content": aggregate_result,
-                "response_uuid": response_uuid,
-                "first_chunk": first_chunk,
-                "last_chunk": last_chunk,
-                "streaming": True,
-            },
-            user_properties=input_message.get_user_properties(),
-        )
-        self.send_to_flow(self.stream_to_flow, message)
-
-    def send_to_next_component(
-        self,
-        input_message,
-        chunk,
-        aggregate_result,
-        response_uuid,
-        first_chunk=False,
-        last_chunk=False,
-    ):
-        message = Message(
-            payload={
-                "chunk": chunk,
-                "content": aggregate_result,
-                "response_uuid": response_uuid,
-                "first_chunk": first_chunk,
-                "last_chunk": last_chunk,
-                "streaming": True,
-            },
-            user_properties=input_message.get_user_properties(),
-        )
-
-        result = {
-            "chunk": chunk,
-            "content": aggregate_result,
-            "response_uuid": response_uuid,
-            "first_chunk": first_chunk,
-            "last_chunk": last_chunk,
-            "streaming": True,
-        }
-
-        self.process_post_invoke(result, message)
+        pass
