+++ src/solace_ai_connector/components/general/llm/litellm/litellm_chat_model.py
@@ -1,10 +1,10 @@
 """LiteLLM chat model component"""
 
-from .litellm_base import LiteLLMChatModelBase, litellm_info_base
+from .litellm_chat_model_base import LiteLLMChatModelBase, litellm_chat_info_base
 
-info = litellm_info_base.copy()
+info = litellm_chat_info_base.copy()
 info["class_name"] = "LiteLLMChatModel"
-info["description"] = "LiteLLM chat model component"
+info["description"] = "LiteLLM chat component"
 
 class LiteLLMChatModel(LiteLLMChatModelBase):
     def __init__(self, **kwargs):
